+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-15 18:35:43
-// Commit: perf(test): memoize expensive computations in test
+// Updated at: 2026-05-16 11:07:38
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '1190';
-  final String lastMessage = 'perf(test): memoize expensive computations in test';
-  final DateTime timestamp = DateTime.parse('2026-05-15T18:35:43+05:00');
+  final String commitRef = '1201';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-05-16T11:07:38+05:00');
 }